+ Finished custom view for sell module

See merge request kkm_android/ofdapp!16

+++ ui/src/main/java/kg/nurtelecom/ofd/sell/CustomCardView.kt
@@ -4,11 +4,11 @@ import android.content.Context
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import androidx.annotation.DimenRes
-import androidx.annotation.Dimension
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import com.google.android.material.card.MaterialCardView
 import kg.nurtelecom.ui.databinding.ViewCardBinding
+import java.math.BigDecimal
 
 class CustomCardView(context: Context, attrs: AttributeSet? = null) :
     MaterialCardView(context, attrs) {
@@ -20,7 +20,7 @@ class CustomCardView(context: Context, attrs: AttributeSet? = null) :
         binding.cardTitle.setText(title)
     }
 
-    fun setCardContent(value: Double) {
+    fun setCardContent(value: BigDecimal) {
         binding.cardContent.text = value.toString()
     }
